+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 * Device addresses now must include the offset (i.e., 40001 is not subtracted).
 
+### 🚀 New
+
+* Add proportional adaptor.
+
 
 ## 0.3.0 - October 12, 2021